+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # OSGL HTTP Change Log
 
 1.6.0
-* Add `CurrentStateStorage` interface to replace previous `Current`
+* Add `incr()` and `decr()` method in `H.Cookie` #20
+* Add `CurrentStateStorage` interface to replace previous `Current` #21
+* By default allow `HttpConfig.xForwardedAllowed` #22
 
 1.5.2 - 02/Apr/2018
 * `H.Request.resolveAcceptFormat()` cannot resolve format for image types #19